Movie_Review_Sentiment_Analysis/deployment/requirements.txt
streamlit
tensorflow-cpu
numpy

